# Conflicts:
#	Smoulder/TemperatureAnalysis/Smoulder/Distributor.cs
#	Smoulder/TemperatureAnalysis/Smoulder/Loader.cs
#	Smoulder/TemperatureAnalysis/Smoulder/OutputDataSegment.cs
#	Smoulder/TemperatureAnalysis/Smoulder/Processor.cs
#	Smoulder/TemperatureAnalysis/TemperatureData.csv

+++ Smoulder/TemperatureAnalysis/Smoulder/LoadedTempData.cs
@@ -7,6 +7,6 @@ namespace TemperatureAnalysis.Smoulder
     {
         public int Id;
         public DateTime Time;
-        public decimal Temperature;
+        public double Temperature;
     }
 }
